rename: context 디렉토리 생성
