+++ src/components/rightSide/RightSide.tsx
@@ -2,7 +2,7 @@ import MyProfile from "./MyProfile";
 import FriendAndDmBar from "./FriendAndDmBar";
 import OtherUserList from "./OtherUserList";
 import { useState } from "react";
-import { UserDropContext } from "../../hooks/UserDropContext";
+import { UserDropContext } from "hooks/context/UserDropContext";
 
 export default function RightSide() {
   const [droppedUser, setDroppedUser] = useState("");
